move ODE, SS, DS to ext
