+++ Project.toml
@@ -6,8 +6,6 @@ authors = ["fbelik <fjbelik@gmail.com> and contributors"]
 [deps]
 ArnoldiMethod = "ec485272-7323-5ecc-a04f-4719b315124d"
 Arpack = "7d9fca2a-8960-54d3-9f78-7d1dccf2cb97"
-ControlSystems = "a6e380b2-a6ca-5380-bf3e-84a91bcd477e"
-DescriptorSystems = "a81e2ce2-54d1-11eb-2c75-db236b00f339"
 HiGHS = "87dc4568-4c63-4d18-b0c0-bb2238e4078b"
 Ipopt = "b6b21f68-93f8-5de0-b562-5493be1d77c9"
 JuMP = "4076af6c-e467-56ae-b986-b466b2749572"
@@ -16,13 +14,23 @@ LinearMaps = "7a12625a-238d-50fd-b39a-03d52299707e"
 MathOptInterface = "b8f27783-ece8-5eb3-8dc8-9495eed66fee"
 MatrixEquations = "99c1a7ee-ab34-5fd5-8076-27c950a045f4"
 NearestNeighbors = "b8a86587-4115-5ab1-83bc-aa920d37bbce"
-OrdinaryDiffEq = "1dea7af3-3e70-54e6-95c3-0bf5283fa5ed"
 Printf = "de0858da-6303-5e67-8744-51eddeeeb8d7"
 ProgressBars = "49802e3a-d2f1-5c88-81d8-b72133a6f568"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
 StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
 UpdatableQRFactorizations = "8b110c62-ada2-4dd7-b53a-29f29fe8f7f4"
+
+[weakdeps]
+ControlSystems = "a6e380b2-a6ca-5380-bf3e-84a91bcd477e"
+DescriptorSystems = "a81e2ce2-54d1-11eb-2c75-db236b00f339"
+OrdinaryDiffEq = "1dea7af3-3e70-54e6-95c3-0bf5283fa5ed"
+
+[extensions]
+ODEExt = "OrdinaryDiffEq"
+ControlSystemsExt = "ControlSystems"
+DescriptorSystemsExt = "DescriptorSystems"
+
 
 [compat]
 ArnoldiMethod = "0.4.0"
