+++ src/models/model.jl
@@ -36,8 +36,8 @@ function (m::NonstationaryModel)(p)
     error("Must implement solution model(p) for NonstationaryModel")
 end
 
-function to_ode_problem(m::NonstationaryModel, p=nothing; x0=0.0, tspan=(0,1))
-    error("Must implement to_ode_problem for NonstationaryModel")
+function to_ode_problem(m::NonstationaryModel, p=nothing; u=(t->0), x0=0.0, tspan=(0,1))
+    error("Must implement to_ode_problem for NonstationaryModel; must import OrdinaryDiffEq to use this method")
 end
 
 function output_type(m::NonstationaryModel)
